refactor: rewrite the core project structure
BREAKING CHANGE: the import way for some module may be changed
