--- packages/core/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/core';
