+++ packages/core/package.json
@@ -24,8 +24,7 @@
     "README.md"
   ],
   "scripts": {
-    "build": "tsc index.ts --target es5 --outDir dist --declaration true",
-    "build2": "tsc --project tsconfig.json",
+    "build": "tsc",
     "test": "jest"
   },
   "dependencies": {
